rename BuHeader and BuResult
