+++ frontend_new/src/pages/SearchBu/SearchBu.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import SearchBar from '../../components/SearchBar';  
 import Footer from '../../components/Footer';
-import Bu from './components/Bu';
-import Result from './components/Result';
+import BuHeader from './components/BuHeader';
+import BuResult from './components/BuResult';
 import { getBuById } from '../../endpoints/bu.api';
 import { useParams } from 'react-router-dom';
 import ManualAutomatic from './components/ManualAutomatic';
@@ -36,9 +36,9 @@ function SearchBu() {
       {id ? (
         <div className='flex place-content-center p-[20px]'>
           <div className='flex-col items-center space-y-[20px]'>     
-            <Bu onSendToMonitor={() => setIsWarningVisible(true)} id={id} />
-            <Result cargo="presidente" />
-            <Result cargo="presidente" />
+            <BuHeader onSendToMonitor={() => setIsWarningVisible(true)} id={id} />
+            <BuResult cargo="presidente" />
+            {/* TODO mostrar outros cargos do BU */}
             {/* <Result cargo="senador" />
             <Result cargo="deputado federal" />
             <Result cargo="deputado estadual" /> */}
