+++ frontend_new/src/pages/SearchBu/components/BuResult.jsx
